Ability to reply wit a predetermined object over websocket for testing
This commit replaces the first pass at an http server function for testing
utility with one that will accept a given route and string body that the
websocket will reply with given a message to that route.

This will allow us to test that the app handles various objects over the wire
as expected.

Signed-off-by: Moses Mendoza <moses@puppet.com>