Remove side-effects in testUtils, cleanup
The methods to obtain a reference to an http server would also start the
server, violating SRP and adding a side effect (though an explicit one). This
revises the methods to instead just return the server without starting,
defering to the caller to start the server. This allows us to do more complex
things like add multiple handlers at different routes.

Signed-off-by: Moses Mendoza <moses@puppet.com>